backend/requirements.txt
# Core Framework
fastapi==0.119.0
uvicorn==0.38.0
pydantic==2.12.3
python-dotenv==1.1.1

# AI & ML
openai>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.2.0
langchain-community>=0.1.0
torch>=2.0.0

# Time Series Forecasting
prophet==1.1.7
pandas==2.3.3
numpy==2.3.4

# Database
pymongo==4.15.3
dnspython==2.8.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Dependencies (auto-installed but listed for clarity)
starlette==0.48.0
pydantic-core==2.41.4
anyio==4.11.0
httpx==0.28.1
requests==2.32.5

